requirements.txt
# AI Education Assistant Requirements
# Install with: pip install -r requirements.txt

# Core Flask framework
Flask==2.3.3
Flask-CORS==4.0.0

# OpenAI API (updated to latest compatible version)
openai>=1.99.0

# Groq API (for alternative AI backend)
groq>=0.13.0

# Markdown to HTML conversion
markdown>=3.7

# PDF processing
PyMuPDF>=1.26.0

# Image processing (for PDF to image conversion)
Pillow>=11.3.0

# Utilities
python-dotenv>=1.1.0
Werkzeug==2.3.7

# Optional: For better logging and monitoring
colorama==0.4.6